requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
python-dotenv>=1.0.0
urllib3>=2.0.0
sqlite-vec>=0.1.0
numpy>=1.24.0